+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.5.0] - Unreleased
 - Added support for scrolling through long command lists.
 - Added visual border around the window.
+- Added option to not show the window border.
 
 ## [0.4.0] - 2021-02-08
 - Added option to not echo commands when executed (-n, --no-echo)